set up an erlang project
I don't really want to deal with concurrent messages and things in C.
Erlang is literally made to handle messages, at multiple levels. At some
point I'll use this C code to make a native GUI for interacting with the
message history that the erlang makes available. I might even use
C+SQLite to make a cache of info that can be displayed without even
starting the erlang node, and get the C GUI to start up the node itself
after that.