reuse closed streams
We use an offset to track what the application thinks the datagram IDs
are.

I had to reintroduce has_control as a boolean, since if you opened a
stream and it is closed now, you have control, but if you didn't open
it, you don't have control, so that's what I want to store and case on
in other parts parts of the logic.